reorganize into maneuvers and planners according to fbouchar/dev
