+++ sv/SDVTrafficState.py
@@ -25,7 +25,7 @@ from Actor import VehicleState
 from mapping.LaneletMap import *
 from SimConfig import *
 from sv.VehicleBase import Vehicle
-from sv.ManeuverConfig import LaneConfig
+from sv.maneuvers.Config import LaneConfig
 from sv.SDVRoute import SDVRoute
 from TickSync import TickSync
 from util.Transformations import (OutsideRefPathException, frenet_to_sim_frame,frenet_to_sim_position, sim_to_frenet_frame,sim_to_frenet_position)
