+++ sv/planners/btree/BehaviorLayer.py
@@ -2,11 +2,11 @@
 #rqueiroz@uwaterloo.ca
 #d43sharm@uwaterloo.ca
 
-from sv.ManeuverConfig import *
-from sv.btree.BTreeInterpreter import *
-from sv.btree.BTreeLeaves import *
-from sv.btree.BTreeConditions import *
-from sv.btree.BTreeRender import *
+from sv.maneuvers.Config import *
+from sv.planners.btree.BTreeInterpreter import *
+from sv.planners.btree.BTreeLeaves import *
+from sv.planners.btree.BTreeConditions import *
+from sv.planners.btree.BTreeRender import *
 
 import logging
 log = logging.getLogger(__name__)
