+++ sv/planners/btree/parser/BTreeDSLLexer.py
@@ -9,7 +9,7 @@ else:
 
 
 from antlr_denter.DenterHelper import DenterHelper
-from sv.btree.parser.BTreeDSLParser import BTreeDSLParser
+from sv.planners.btree.parser.BTreeDSLParser import BTreeDSLParser
